Improve FlooNoC FPGA setup
A HLS traffic generator has been added to the FPGA build flow.
This becomes part of the implemented FlooNoC setup which is built by
"floo_fpga_build.tcl".
Besides, the generated components have been improved, solving an error
in the FlooNoC top-level when instantiating the tiles.
The error arose because x-y for loops were inverted, thus breaking the tile mapping.